Move `uiData` into `index`
